grammar integration step - prepared lang
